v2.4 wip

+++ api/include/wsa_lib.h
@@ -237,7 +237,7 @@ int16_t wsa_read_iq_packet_raw_matlab(struct wsa_device* const device,
 		uint8_t* const data_buffer, 
 		const uint16_t samples_per_packet, uint8_t* context_present,
 	int32_t* rec_indicator_field, int32_t* rec_reference_point, int64_t* rec_frequency, int16_t* rec_gain_if, int16_t* rec_gain_rf,	int32_t* rec_temperature,
-	int32_t* dig_indicator_field, long double* dig_bandwidth, double* dig_reference_level, long double* digrf_frequency_offset);		
+	int32_t* dig_indicator_field, int64_t* dig_bandwidth, int32_t* dig_reference_level, int64_t* digrf_frequency_offset);
